+++ src/class/BoostDetector.ts
@@ -1,0 +1,37 @@
+import { Client, GuildMember } from 'discord.js';
+import { EventEmitter } from 'events';
+import { DJSError, errorkeys } from '../error';
+
+interface BoostDetectorEvents {
+    'boostCreate': [member: GuildMember]
+    'boostRemove': [member: GuildMember]
+};
+
+export class BoostDetector extends EventEmitter {
+    constructor(client: Client) {
+        if (!client) throw new DJSError(errorkeys.MissingParam);
+
+        if (!(client instanceof Client)) throw new DJSError(errorkeys.TypeError, 'Client');
+
+        super({
+            captureRejections: true
+        });
+
+        client.on('guildMemberUpdate', (oldM, newM) => {
+            if (newM.premiumSinceTimestamp != oldM.premiumSinceTimestamp) {
+                this.emit('boostRemove', newM);
+
+                return;
+            };
+            
+            if (oldM.premiumSinceTimestamp != newM.premiumSinceTimestamp) {
+                this.emit('boostCreate', newM);
+
+                return;
+            };
+        });
+    };
+
+    public on<K extends keyof BoostDetectorEvents>(eventName: K, listener: (...args: BoostDetectorEvents[K]) => void) { return this };
+    public once<K extends keyof BoostDetectorEvents>(eventName: K, listener: (...args: BoostDetectorEvents[K]) => void) { return this };
+};