+++ src/class/index.ts
@@ -2,3 +2,4 @@ export * from './ApplicationCommandsLoader';
 export * from './FileBuilder';
 export * from './JSONDatabase';
 export * from './StringSelectMenuPaginatorBuilder';
+export * from './BoostDetector';