+++ src/func/index.ts
@@ -3,3 +3,5 @@ export * from './createDiscordTimestamp';
 export * from './isDiscordInviteURL';
 export * from './isWebURL';
 export * from './reverseString';
+export * from './sleep';
+export * from './randomizedString';